pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gis-bootcamp"
version = "0.1.0"
description = "Production-grade GIS tools for data engineers"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "geopandas>=0.14.0",
    "shapely>=2.0.0",
    "pyproj>=3.6.0",
    "rasterio>=1.3.0",
    "pyarrow>=14.0.0",
    "geopy>=2.4.0",
    "requests>=2.31.0",
]



[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1",
]
